+++ src/aks-preview/azext_aks_preview/custom.py
@@ -1162,7 +1162,9 @@ def aks_create(
     enable_gateway_api=False,
     # app routing istio
     enable_app_routing_istio=False,
-    enable_hosted_system=False
+    enable_hosted_system=False,
+    # health monitor
+    enable_continuous_control_plane_and_addon_monitor=False,
 ):
     # DO NOT MOVE: get all the original parameters and save them as a dictionary
     raw_parameters = locals()
@@ -1409,6 +1411,9 @@ def aks_update(
     # application load balancer
     enable_application_load_balancer=False,
     disable_application_load_balancer=False,
+    # health monitor
+    enable_continuous_control_plane_and_addon_monitor=False,
+    disable_continuous_control_plane_and_addon_monitor=False,
 ):
     # DO NOT MOVE: get all the original parameters and save them as a dictionary
     raw_parameters = locals()